Move UIView extension to its own file
